improve: generalize block finder interface for EVM
