+++ src/arch/evm/index.ts
@@ -1,1 +1,2 @@
 export * from "./SpokeUtils";
+export * from "./BlockUtils";
